monitor-config-manager: fix implicit conversion warning

+++ backends/gf-monitor-config-manager.c
@@ -598,7 +598,7 @@ assign_monitor_crtcs (GfMonitorManager        *manager,
       g_set_error (error, G_IO_ERROR, G_IO_ERROR_FAILED,
                    "Invalid mode %dx%d (%f) for monitor '%s %s'",
                    monitor_mode_spec->width, monitor_mode_spec->height,
-                   monitor_mode_spec->refresh_rate,
+                   (gdouble) monitor_mode_spec->refresh_rate,
                    monitor_spec->vendor, monitor_spec->product);
 
       return FALSE;
